+++ staking/staking.go
@@ -19,8 +19,9 @@ func New(
 	witness Witness,
 	ethClient AllEthereumClient,
 	netp *netparams.Store,
+	evtFwd EvtForwarder,
 ) (*Accounting, *StakeVerifier) {
-	accs := NewAccounting(log, cfg, broker, ethClient)
+	accs := NewAccounting(log, cfg, broker, ethClient, evtFwd, witness, tt)
 	ethCfns := NewEthereumConfirmations(ethClient, nil)
 	ocv := NewOnChainVerifier(cfg, log, ethClient, ethCfns)
 	sakeV := NewStakeVerifier(log, cfg, accs, tt, witness, broker, ocv)