+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - [4514](https://github.com/vegaprotocol/vega/pull/4514) - Add network limits service and events
 - [4516](https://github.com/vegaprotocol/vega/pull/4516) - Add a command to cleanup all vega node state
 - [4531](https://github.com/vegaprotocol/vega/pull/4531) - Remove Float from network parameters, use `num.Decimal` instead
+- [4537](https://github.com/vegaprotocol/vega/pull/4537) - Send staking asset total supply through consensus
 
 
 ### 🐛 Fixes